Move watch manager screens to settings too
