+++ mobile/src/main/kotlin/com/boswelja/smartwatchextensions/MainApplication.kt
@@ -52,7 +52,6 @@ class MainApplication : Application() {
 
             modules(
                 dashboardModule,
-                com.boswelja.smartwatchextensions.devicemanagement.watchManagerModule,
                 mainModule,
                 phoneLockingModule,
                 appSettingsModule,
