+++ mobile/src/main/kotlin/com/boswelja/smartwatchextensions/settings/DI.kt
@@ -2,6 +2,8 @@ package com.boswelja.smartwatchextensions.settings
 
 import android.content.Context
 import com.boswelja.smartwatchextensions.core.settings.appSettingsStore
+import com.boswelja.smartwatchextensions.settings.ui.WatchManagerViewModel
+import com.boswelja.smartwatchextensions.settings.ui.register.RegisterWatchViewModel
 import com.boswelja.smartwatchextensions.settings.ui.AppSettingsViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
@@ -11,4 +13,6 @@ import org.koin.dsl.module
  */
 val appSettingsModule = module {
     viewModel { AppSettingsViewModel(get(), get(), get<Context>().appSettingsStore) }
+    viewModel { WatchManagerViewModel(get(), get()) }
+    viewModel { RegisterWatchViewModel(get()) }
 }
