+++ mobile/src/main/kotlin/com/boswelja/smartwatchextensions/settings/ui/SettingsDestination.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.activity
 import androidx.navigation.compose.composable
-import com.boswelja.smartwatchextensions.devicemanagement.ui.watchManagerGraph
+import com.boswelja.smartwatchextensions.settings.ui.watchManagerGraph
 
 /**
  * All destinations reachable from withing App Settings.
